separate components into ui folder
