+++ frontend/components/ui/Header.js
@@ -1,5 +1,5 @@
 import styles from "./Header.module.css";
-import Navbar from "@/components/Navbar";
+import Navbar from "@/components/ui/Navbar";
 
 export default function Header() {
   return (
